backend/requirements.txt
boto3==1.40.54
fastapi==0.119.0
geopy==2.4.1
pydantic==2.12.3
python-dotenv==1.1.1
Requests==2.32.5
streamlit==1.50.0
